+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     * `glean.restarted` events will be included in Custom Pings with other events to rationalize event timestamps across restarts.
   * `test_reset_glean` will remove all previous data if asked to clear stores, even if Glean never has been initialized ([#2294](https://github.com/mozilla/glean/pull/2294))
   * Upgrade to `glean_parser` v6.5.0, with support for `Cow` in Rust code ([#2300](https://github.com/mozilla/glean/issues/2300))
+  * API REMOVED: The deprecated-since-v38 `event` metric `record(map)` API has been removed ([bug 1802550](https://bugzilla.mozilla.org/show_bug.cgi?id=1802550))
 
 # v51.8.3 (2022-11-25)